# WebClient/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Service",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Loans": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "GetLoans",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Loans"
        ],
        "operationId": "AddLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Loans"
        ],
        "operationId": "UpdateLoan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestModel"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestModel"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoanRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Loans/{id}": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "GetLoan",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Loan"
                }
              }
            }
          }
        }
      }
    },
    "/api/Loans/Search": {
      "get": {
        "tags": [
          "Loans"
        ],
        "operationId": "SearchLoans",
        "parameters": [
          {
            "name": "bookId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "identificationNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Loan"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Author": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "firstname": {
            "type": "string",
            "nullable": true
          },
          "lastname": {
            "type": "string",
            "nullable": true
          },
          "birthdate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "pseudonym": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "year": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "publisher": {
            "type": "string",
            "nullable": true
          },
          "isbn": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "author": {
            "$ref": "#/components/schemas/Author"
          },
          "genres": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Genre"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Genre": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Loan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "identificationNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "loanedOn": {
            "type": "string",
            "format": "date-time"
          },
          "book": {
            "$ref": "#/components/schemas/Book"
          }
        },
        "additionalProperties": false
      },
      "LoanRequestModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "bookId": {
            "type": "integer",
            "format": "int32"
          },
          "identificationNumber": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "loanedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}